zlib: refactor to use files
We need to have a position indicator when reading files. Files have
one built-in, which allows us to not have to keep track of offsets.

In doing this, we actually don't lose any generality. File-like objects
include streams. You can take any bytes-like object and wrap it in
io.BytesIO to get a file-like object for that buffer (which we do here).